requirements.txt
streamlit>=1.32          
torch>=2.2.0            
torchvision>=0.17       
Pillow>=10.0           
numpy>=1.26              
matplotlib>=3.9         
